add discussions to topics page
